+++ backend/src/main/java/com/woowacourse/ternoko/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.controller;
 
 import com.woowacourse.ternoko.domain.AvailableDateTime;
-import com.woowacourse.ternoko.dto.AvailableDateTimesRequest;
+import com.woowacourse.ternoko.dto.request.AvailableDateTimesRequest;
 import com.woowacourse.ternoko.dto.AvailableDateTimesResponse;
 import com.woowacourse.ternoko.dto.CoachesResponse;
 import com.woowacourse.ternoko.service.MemberService;
