+++ backend/src/main/java/com/woowacourse/ternoko/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.dto;
+package com.woowacourse.ternoko.dto.request;
 
 import java.time.LocalDateTime;
 import java.util.List;
