+++ backend/src/main/java/com/woowacourse/ternoko/service/MemberService.java
@@ -3,7 +3,7 @@ package com.woowacourse.ternoko.service;
 import com.woowacourse.ternoko.domain.AvailableDateTime;
 import com.woowacourse.ternoko.domain.Member;
 import com.woowacourse.ternoko.domain.Type;
-import com.woowacourse.ternoko.dto.AvailableDateTimesRequest;
+import com.woowacourse.ternoko.dto.request.AvailableDateTimesRequest;
 import com.woowacourse.ternoko.dto.CoachResponse;
 import com.woowacourse.ternoko.dto.CoachesResponse;
 import com.woowacourse.ternoko.repository.AvailableDateTimeRepository;
