+++ backend/src/main/java/com/woowacourse/ternoko/service/ReservationService.java
@@ -4,9 +4,8 @@ import com.woowacourse.ternoko.domain.FormItem;
 import com.woowacourse.ternoko.domain.Interview;
 import com.woowacourse.ternoko.domain.Member;
 import com.woowacourse.ternoko.domain.Reservation;
-import com.woowacourse.ternoko.dto.FormItemRequest;
-import com.woowacourse.ternoko.dto.FormItemResponse;
-import com.woowacourse.ternoko.dto.ReservationRequest;
+import com.woowacourse.ternoko.dto.request.FormItemRequest;
+import com.woowacourse.ternoko.dto.request.ReservationRequest;
 import com.woowacourse.ternoko.dto.ReservationResponse;
 import com.woowacourse.ternoko.dto.ScheduleResponse;
 import com.woowacourse.ternoko.repository.FormItemRepository;
