+++ backend/src/test/java/com/woowacourse/ternoko/service/MemberServiceTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.ternoko.domain.AvailableDateTime;
-import com.woowacourse.ternoko.dto.AvailableDateTimesRequest;
+import com.woowacourse.ternoko.dto.request.AvailableDateTimesRequest;
 import com.woowacourse.ternoko.dto.CoachesResponse;
 import java.time.LocalDateTime;
 import java.util.List;
